refactor: irgen
