+++ internal/compiler/irgen/irgen.go
@@ -2,7 +2,6 @@ package irgen
 
 import (
 	"errors"
-	"fmt"
 
 	"github.com/nevalang/neva/internal/compiler"
 	src "github.com/nevalang/neva/internal/compiler/sourcecode"
@@ -16,9 +15,9 @@ type Generator struct{}
 
 type (
 	nodeContext struct {
-		path       []string   // Path to current node including current node
-		node       src.Node   // Node definition
-		portsUsage portsUsage // How parent network uses this node's ports
+		path       []string
+		node       src.Node
+		portsUsage portsUsage
 	}
 
 	portsUsage struct {
@@ -74,127 +73,6 @@ func (g Generator) Generate(
 	}
 
 	return result, nil
-}
-
-func (g Generator) processComponentNode(
-	nodeCtx nodeContext,
-	scope src.Scope,
-	result *ir.Program,
-) *compiler.Error {
-	ref := nodeCtx.node.EntityRef
-
-	entity, found, err := scope.Entity(ref)
-	if err != nil {
-		return &compiler.Error{
-			Err:      err,
-			Location: &scope.Location,
-		}
-	}
-
-	comp := entity.Component
-	in := g.insertAndReturnInports(nodeCtx, result)   // for inports we only use parent context because all inports are used
-	out := g.insertAndReturnOutports(nodeCtx, result) //  for outports we use both parent context and component's interface
-
-	funcRef, err := getFuncRef(
-		comp,
-		nodeCtx.node.TypeArgs,
-	)
-	if err != nil {
-		return &compiler.Error{
-			Err:      err,
-			Location: &found,
-			Meta:     &comp.Meta,
-		}
-	}
-
-	if funcRef != "" {
-		funcCall, err := getFuncCall(nodeCtx, scope, funcRef, in, out)
-		if err != nil {
-			return err
-		}
-		result.Funcs = append(result.Funcs, funcCall)
-		return nil
-	}
-
-	newScope := scope.WithLocation(found)
-
-	subnodesPortsUsage, err := g.processNetwork(comp.Net, nodeCtx, result)
-	if err != nil {
-		return &compiler.Error{
-			Err:      err,
-			Location: &newScope.Location,
-		}
-	}
-
-	for nodeName, node := range comp.Nodes {
-		nodePortsUsage := subnodesPortsUsage[nodeName]
-
-		subNodeCtx := nodeContext{
-			path:       append(nodeCtx.path, nodeName),
-			portsUsage: nodePortsUsage,
-			node:       node,
-		}
-
-		scopeToUse := getSubnodeScope(
-			nodeCtx.node.Deps,
-			nodeName,
-			subNodeCtx,
-			scope,
-			newScope,
-		)
-
-		if err := g.processComponentNode(
-			subNodeCtx,
-			scopeToUse,
-			result,
-		); err != nil {
-			return &compiler.Error{
-				Err:      fmt.Errorf("%w: node '%v'", err, nodeName),
-				Location: &found,
-				Meta:     &comp.Meta,
-			}
-		}
-	}
-
-	return nil
-}
-
-func getSubnodeScope(
-	deps map[string]src.Node,
-	nodeName string,
-	subNodeCtx nodeContext,
-	scope src.Scope,
-	newScope src.Scope,
-) src.Scope {
-	var scopeToUseThisTime src.Scope
-	if dep, ok := deps[nodeName]; ok { // is interface node
-		subNodeCtx.node = dep
-		scopeToUseThisTime = scope
-	} else {
-		scopeToUseThisTime = newScope
-	}
-	return scopeToUseThisTime
-}
-
-func getFuncCall(
-	nodeCtx nodeContext,
-	scope src.Scope,
-	funcRef string,
-	in []ir.PortAddr,
-	out []ir.PortAddr,
-) (ir.FuncCall, *compiler.Error) {
-	cfgMsg, err := getCfgMsg(nodeCtx.node, scope)
-	if err != nil {
-		return ir.FuncCall{}, &compiler.Error{
-			Err:      err,
-			Location: &scope.Location,
-		}
-	}
-	return ir.FuncCall{
-		Ref: funcRef,
-		IO:  ir.FuncIO{In: in, Out: out},
-		Msg: cfgMsg,
-	}, nil
 }
 
 func New() Generator {
